requirements.txt
google-analytics-data
pandas
boto3